+++ xdef/src/main/java/org/xdef/XDCurrency.java
@@ -1,0 +1,6 @@
+package org.xdef;
+
+/** Currency value.
+ * @author Vaclav Trojan
+ */
+public interface XDCurrency extends XDValue {}